requirements.txt
flask
nltk
spacy
scikit-learn
pymongo
PyPDF2
python-docx
pandas
requests
requests-oauthlib
